requirements.txt
django
djangorestframework
geopy
pillow